refactor: decompose Methods.java into focused utility classes
Methods.java was 1,790 lines mixing 7+ concerns. Extract 5 new classes:
- HashUtils: consolidate 3 duplicate MD5 implementations
- YamlUtils: YAML reading and manipulation
- ArchiveDownloader: curl/zip/tar extraction and S3 downloads
- FileTraversalUtils: directory walking and file collection
- NextflowOutputFilter: output line filtering and sanitization

Methods.java is now a 488-line thin delegation facade.
Delete duplicate filterNextflowOutput Groovy adapter.
Update NfCoreUtils and CsvUtils to use HashUtils.
